misc: rename thrift remote implementations
