+++ velox/functions/remote/utils/CMakeLists.txt
@@ -16,4 +16,4 @@ add_library(velox_functions_remote_utils RemoteFunctionServiceProvider.cpp)
 
 target_link_libraries(
   velox_functions_remote_utils
-  PUBLIC velox_functions_remote_server)
+  PUBLIC velox_functions_remote_thrift_server)
